some refactoring, index packing

+++ src/bdd.rs
@@ -4,9 +4,9 @@ use std::io::prelude::*;
 
 pub type VarLabel = u16;
 /// Index into BDD table
-pub type TableIndex = u16;
+pub type TableIndex = u32;
 
-pub type RefCount = u8;
+pub type RefCount = u16;
 
 /// A BDD pointer
 #[derive(Debug, Hash, Clone, PartialEq, Eq)]
